delete the old repr
